+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Elliptic curve TLS certificates are now supported in pcsd ([ghissue#123])
+- Support for corosync option `totem.block_unlisted_ips` ([rhbz#1720221])
 
 ### Fixed
 - Node attribute expressions are now correctly reported as not allowed in
@@ -13,6 +14,7 @@
 - python3-openssl was replaced with python3-cryptography ([rhbz#1927404])
 
 [ghissue#123]: https://github.com/ClusterLabs/pcs/issues/123
+[rhbz#1720221]: https://bugzilla.redhat.com/show_bug.cgi?id=1720221
 [rhbz#1896458]: https://bugzilla.redhat.com/show_bug.cgi?id=1896458
 [rhbz#1927404]: https://bugzilla.redhat.com/show_bug.cgi?id=1927404